// crud/src/app/enums/marthon-enums.enum.ts
export enum RacesEnum {
    Trka5km = 5,
    Trka10km = 10,
    Trka21km = 21,
    Trka41km = 41
}


export enum BloodType {
    A_POSITIVE = "A+",
    A_NEGATIVE = "A-",
    B_POSITIVE = "B+",
    B_NEGATIVE = "B-",
    AB_POSITIVE = "AB+",
    AB_NEGATIVE = "AB-",
    O_POSITIVE = "O+",
    O_NEGATIVE = "O-",
  };